Fix Show more button in leaderboard widget
