+++ frontend/src/modules/widget/components/v2/activity/widget-total-activities.vue
@@ -22,7 +22,7 @@
         <template #default="{ resultSet, loading, error }">
           <!-- Loading -->
           <app-widget-loading
-            v-if="loading"
+            v-if="loading || error === undefined"
             size="small"
           />
